+++ src/take_measurements.h
@@ -74,7 +74,7 @@ bool isItSafeToCharge();                               // See if it is safe to c
  * @details Provides data on the signal strength and quality
  * 
  */
-void getSignalStrength();
+void logSignalStrength();
 
 /**
  * @brief This function is called once a hardware interrupt is triggered by the device's sensor